Port generic TFRU fixes with 56-commit audit

+++ src/main/java/zmaster587/advancedRocketry/inventory/GuiOreMappingSatellite.java
@@ -335,7 +335,7 @@ public class GuiOreMappingSatellite extends GuiContainer {
 		//Render player location
 		float offsetX = playerPosX - xCenter;
 		float offsetY = zCenter - playerPosZ ;
-		double numPixels = SCREEN_SIZE/scanSize;//(scanSize/(float)(SCREEN_SIZE*radius));
+		double numPixels = (double)SCREEN_SIZE/scanSize;//(scanSize/(float)(SCREEN_SIZE*radius));
 
 
 		float radius = 2;
